Moved the clean test class to the `TryAtSoftware.CleanTests.Core` project. Improved the test discovery process whenever clean and standard tests cases are being mixed
