--- TryAtSoftware.CleanTests.Core/Attributes/TestSuiteAttribute.cs
@@ -1,8 +1,0 @@
-﻿namespace TryAtSoftware.CleanTests.Core.Attributes;
-
-using System;
-
-[AttributeUsage(AttributeTargets.Class)]
-public class TestSuiteAttribute : Attribute
-{
-}
