+++ TryAtSoftware.CleanTests.Core/TryAtSoftware.CleanTests.Core.csproj
@@ -23,11 +23,7 @@
         <PackageReference Include="System.Threading.Tasks.Dataflow" Version="7.0.0" />
         <PackageReference Include="TryAtSoftware.Extensions.Collections" Version="1.0.1" />
         <PackageReference Include="TryAtSoftware.Extensions.Reflection" Version="1.0.1" />
-        <PackageReference Include="xunit.core" Version="2.4.2" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <Folder Include="XUnit\Data" />
+        <PackageReference Include="xunit" Version="2.4.2" />
     </ItemGroup>
 
 </Project>
