+++ TryAtSoftware.CleanTests.Sample/GenericTest.cs
@@ -1,9 +1,11 @@
 ﻿namespace TryAtSoftware.CleanTests.Sample;
 
+using TryAtSoftware.CleanTests.Core;
 using TryAtSoftware.CleanTests.Core.Attributes;
 using TryAtSoftware.CleanTests.Sample.Attributes;
 using Xunit.Abstractions;
 
+[TestSuiteGenericTypeMapping(typeof(NumericAttribute), typeof(int))]
 public class GenericTest<[Numeric] T> : CleanTest
     where T : notnull
 {
